+++ package.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0"?>
 <?xml-model href="http://download.ros.org/schema/package_format3.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
 <package format="3">
-  <name>bno08x_ros2_driver</name>
+  <name>bno08x_driver</name>
   <version>0.0.0</version>
   <description>ROS 2 Driver for CEVA BNO08x IMU</description>
   <maintainer email="bnbhat311@gmail.com">Balachandra Bhat</maintainer>
@@ -12,6 +12,7 @@
   <depend>rclcpp</depend>
   <depend>sensor_msgs</depend>
   <depend>std_msgs</depend>
+  <depend>bno08x_interfaces</depend>
 
   <test_depend>ament_lint_auto</test_depend>
   <test_depend>ament_lint_common</test_depend>
